@@ -1,24 +1,23 @@
 ext.versions = [
-        code           : 7,
-        name           : '1.1.4',
+        code           : 14,
+        name           : '1.2.0',
 
-        minSdk         : 7,
-        targetSdk      : 23,
-        compileSdk     : 23,
-        buildTools     : '23.0.3',
+        minSdk         : 14,
+        targetSdk      : 28,
+        compileSdk     : 28,
+        buildTools     : '28.0.3',
 
         // Library versions
-        supportLibs    : '23.3.0',
         nineOldAndroids: '2.4.0',
         materialicons  : '1.0.1',
         stickyheaders  : '0.5.0'
 ]
 
 ext.libraries = [
-        supportRecyclerView: "com.android.support:recyclerview-v7:$versions.supportLibs",
-        supportAppCompat   : "com.android.support:appcompat-v7:$versions.supportLibs",
-        supportDesign      : "com.android.support:design:$versions.supportLibs",
-        supportCardView    : "com.android.support:cardview-v7:$versions.supportLibs",
+        supportRecyclerView: 'androidx.recyclerview:recyclerview:1.0.0',
+        supportAppCompat   : 'androidx.appcompat:appcompat:1.0.0',
+        supportDesign      : 'com.google.android.material:material:1.0.0',
+        supportCardView    : 'androidx.cardview:cardview:1.0.0',
         supportAnnotations : "com.android.support:support-annotations:$versions.supportLibs",
 
         nineOldAndroids    : "com.nineoldandroids:library:$versions.nineOldAndroids",
